+++ src/Resources/contao/languages/de/modules.php
@@ -8,16 +8,11 @@
  */
 
 /*
- * Backend-Bereich
- */
-$GLOBALS['TL_LANG']['MOD']['bsv'] = array
-(
-	'Berliner Schachverband',
-	'Berliner Schachverband'
-);
-
-/*
  * Backend-Modul
+ *
+ * Das Label des Backend-Bereichs "bsv" wird nicht mehr von diesem Bundle
+ * bereitgestellt, sondern von einer eigenen Erweiterung des Berliner
+ * Schachverbands.
  */
 $GLOBALS['TL_LANG']['MOD']['grandprix'][0] = 'Grand Prix';
 $GLOBALS['TL_LANG']['MOD']['grandprix'][1] = 'Berliner Schnellschach-Grand-Prix-Serie';